+++ CHANGES.md
@@ -6,6 +6,10 @@
 
 - Added support for the `KHR_materials_variants` extension to the glTF reader and writer.
 
+##### Fixes :wrench:
+
+- Fixed a bug that caused meshes to be missing entirely when upsampled from a parent with `UNSIGNED_BYTE` indices.
+
 ### v0.24.0 - 2023-05-01
 
 ##### Additions :tada: